// src/types/game.ts
import type { RowCount } from "../constants/game";

import PlinkoEngine from "../components/Plinko/PlinkoEngine";

export enum BetMode {
  MANUAL = "MANUAL",
  AUTO = "AUTO",
}

/**
 * Game's risk level, which controls the volatility of payout.
 */
export enum RiskLevel {
  LOW = "LOW",
  MEDIUM = "MEDIUM",
  HIGH = "HIGH",
}

/**
 * A record of the bet amount associated to every existing ball in the game
 * that is still in motion.
 *
 * When a ball enters a bin, its record is removed.
 */
export type BetAmountOfExistingBalls = {
  [ballId: number]: number;
};

export type WinRecord = {
  /**
   * UUID of the win record.
   */
  id: string;
  /**
   * How much the player has bet.
   */
  betAmount: number;
  /**
   * Number of pin rows at the time of winning.
   */
  rowCount: RowCount;
  /**
   * Zero-based index of which bin the ball fell into (leftmost bin is 0).
   */
  binIndex: number;
  payout: {
    /**
     * Multiplier for the payout (e.g. `0.3`, `1.5`).
     */
    multiplier: number;
    /**
     * Actual payout amount.
     */
    value: number;
  };
  /**
   * Payout value minus the bet amount.
   */
  profit: number;
};

export type GameState = {
  plinkoEngine: PlinkoEngine | null;
  betAmount: number;
  betAmountOfExistingBalls: BetAmountOfExistingBalls;
  rowCount: RowCount;
  riskLevel: RiskLevel;
  winRecords: WinRecord[];
  totalProfitHistory: number[];
  balance: number;
  binColors: {
    background: string[];
    shadow: string[];
  };
};